исправления после ревью

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -14,5 +14,5 @@ public class User {
     private String login;
     private String name;
     private LocalDate birthday;
-    private Set<Long> friends = new HashSet<>();
+    private Set<Integer> friends = new HashSet<>();
 }
